requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
